@@ -1,13 +1,12 @@
 import React from "react";
 import Login from "../Login";
 import {LABELS_AND_HEADINGS, TEXTS} from "../../helpers/constants";
-import {BadgeCheckIcon, ChevronDoubleDownIcon, ClipboardListIcon, CollectionIcon, UserAddIcon} from "@heroicons/react/solid";
+import {BadgeCheckIcon, ClipboardListIcon, CollectionIcon, UserAddIcon} from "@heroicons/react/solid";
 import shieldWhite from "../../assets/images/shield__white.svg";
 
 export const HeroHeader = () => {
     return (
         <div className={"sms-hero-header mb-5"}>
-
             <div className={"container-fluid"}>
                 <div className={"row"}>
                     <div className={"col-12 col-xl-6 p-5 bg-whale d-flex align-items-center flex-column justify-content-center"}>
@@ -18,24 +17,17 @@ export const HeroHeader = () => {
                                     {LABELS_AND_HEADINGS.SVENSKA_MARVELSAMLARE}
                                 </p>
                             </div>
-                            <div className={"align-self-center mb-4 d-flex align-items-center flex-column"}>
-                                <h2 className={"fs-1 text-primary mb-4"} id={"log-in-section"}>{LABELS_AND_HEADINGS.LOG_IN}</h2>
-                                <ChevronDoubleDownIcon className="sms-icon--large text-info"/>
-                            </div>
                             <Login/>
                         </div>
-                        <h3 className={"mb-3 text-primary fs-large"}>Är du inte registrerad?</h3>
-                        <a href={"#create-account-section"} className={"btn btn-secondary btn-cta mb-5"}>
-                            <UserAddIcon className={"sms-icon"}/> {LABELS_AND_HEADINGS.CREATE_ACCOUNT}
+                        <a href={"#create-account-section"} className={"btn btn-primary btn-cta mb-5"}>
+                            <UserAddIcon className={"sms-icon"}/> {LABELS_AND_HEADINGS.CREATE_ACCOUNT_CTA}
                         </a>
                     </div>
-
                     <div className={"col-12 col-xl-6 p-5 d-flex align-items-center justify-content-center flex-column text-center"}>
                         <h1 className={"text-primary"}>
                             {LABELS_AND_HEADINGS.WELCOME}
                         </h1>
                         <p className={"lead mb-5"}>{TEXTS.DO_YOU_COLLECT}</p>
-
                         <div className={"mb-4"}>
                             <BadgeCheckIcon className="sms-icon--large mb-2 text-info"/>
                             <p>{TEXTS.MANAGE_YOUR_COLLECTION}</p>
@@ -49,7 +41,6 @@ export const HeroHeader = () => {
                             <p>{TEXTS.ALWAYS_AVAILABLE}</p>
                         </div>
                     </div>
-
                 </div>
             </div>
         </div>
