@@ -3,12 +3,14 @@ import {Link} from "react-router-dom";
 import {BUCKETS, LABELS_AND_HEADINGS, ROUTES, TABLES} from "../../../helpers/constants";
 import {PencilAltIcon, XCircleIcon} from "@heroicons/react/solid";
 import {handleDelete} from "../../serviceFunctions";
+import {useAppContext} from "../../../context/AppContext";
 
 
 export const PublishersListAdminToolBox = ({publisher, setPublishersData, publishersData}) => {
 
     const editText = LABELS_AND_HEADINGS.EDIT + " " + publisher.name;
     const deleteText = LABELS_AND_HEADINGS.DELETE + " " + publisher.name;
+    const {setInformationMessage} = useAppContext();
 
     return (
         <div className={"ms-2 d-inline-block"}>
@@ -19,7 +21,7 @@ export const PublishersListAdminToolBox = ({publisher, setPublishersData, publis
             <button
                 className={"btn text-danger sms-icon-btn"}
                 aria-label={deleteText}
-                onClick={() => handleDelete(TABLES.PUBLISHERS, publisher.id, publisher.name, setPublishersData, publishersData, publisher.image, BUCKETS.PUBLISHER_IMAGES)}>
+                onClick={() => handleDelete(TABLES.PUBLISHERS, publisher.id, publisher.name, setPublishersData, publishersData, publisher.image, BUCKETS.PUBLISHER_IMAGES, setInformationMessage)}>
                 <XCircleIcon className={"sms-icon--text-lg m-0"}/>
             </button>
         </div>
