@@ -1,8 +1,8 @@
 import React from "react";
 
 
-export const LogMe = ({logThis}) => {
-    console.log("LOG", logThis)
+export const LogMe = ({logThis, logKey}) => {
+    console.log("LOG " + logKey + ": ", logThis)
     console.log("-------")
     return (
         <></>
