@@ -12,10 +12,10 @@ const SignupSuccess = () => {
                     <div className={"align-self-center mb-4 d-flex align-items-center flex-column w-100"}>
                         <h1 className={"text-primary text-center"}>Registreringen lyckades!</h1>
                         <p className={"lead mb-4"}>Nu kan du logga in.</p>
-                        <BadgeCheckIcon className="sms-icon--large text-info"/>
+                        <BadgeCheckIcon className="sms-icon--large text-info mb-4"/>
                         <Login/>
                     </div>
-                    <Link to="/" className={"btn btn-secondary btn-lg mb-5"}>
+                    <Link to="/" className={"btn btn-primary btn-cta mb-5"}>
                         TILLBAKA
                     </Link>
                 </div>
