@@ -2,7 +2,7 @@ import React from "react";
 import {LABELS_AND_HEADINGS, TEXTS} from "../../../helpers/constants";
 import {TitlesCard} from "../../dashboard/dashboardCards/admin/TitlesCard";
 import {PublishersCard} from "../../dashboard/dashboardCards/admin/PublishersCard";
-import {AdminH1} from "../../headings";
+import {HeadingWithBreadCrumbs} from "../../headings";
 import {MailIcon} from "@heroicons/react/solid";
 
 
@@ -12,7 +12,7 @@ export const Admin = () => {
         <main className={"container-fluid main-container"}>
             <div className={"row row-padding--main"}>
                 <div className={"col-12 col-lg-8"}>
-                    <AdminH1 text={LABELS_AND_HEADINGS.ADMIN}/>
+                    <HeadingWithBreadCrumbs text={LABELS_AND_HEADINGS.ADMIN}/>
                     <p className={"lead"}>{TEXTS.ADMIN_LEAD}</p>
                     <p>{TEXTS.ADMIN_INFO}</p>
                     <p><MailIcon className={"sms-icon--text-md"}/> <a href={"mailto: admin@svenskamarvelsamlare.se"}>admin@svenskamarvelsamlare.se</a></p>
