@@ -1,12 +1,13 @@
 import React, {useEffect, useState} from "react";
 import {CLASSES, LABELS_AND_HEADINGS} from "../../../../helpers/constants";
 import {addPublisherData, handleInput} from "../../../serviceFunctions";
-import {handleBacking, hideAndResetMessage, printOptions} from "../../../../helpers/functions";
+import {handleBacking, printOptions} from "../../../../helpers/functions";
 import countryData from "../../../../helpers/valueLists/countries.json";
 import {useCommonFormStates} from "../../../../helpers/customHooks/useCommonFormStates";
-import {AdminH1} from "../../../headings";
+import {HeadingWithBreadCrumbs} from "../../../headings";
 import {ArrowLeftButton} from "../../../minis/ArrowLeftButton";
 import {useNavigate} from "react-router-dom";
+import {useAppContext} from "../../../../context/AppContext";
 
 
 export const AdminPublisherAdd = () => {
@@ -14,36 +15,34 @@ export const AdminPublisherAdd = () => {
     const [
         name, setName,
         description, setDescription,
-        formMessage, setFormMessage,
         formInputClass, setFormInputClass
     ] = useCommonFormStates();
 
+    const {setInformationMessage} = useAppContext();
     const navigate = useNavigate();
-
-    const [countryId, setCountryId] = useState("");
+    const [country_id, setCountry_id] = useState("");
 
     const resetAddPublisherForm = async () => {
         setName("");
         setDescription("");
         setFormInputClass(CLASSES.FORM_INPUT_ERROR);
-        hideAndResetMessage(setFormMessage);
     }
 
     useEffect(() => {
-        if (countryId && name !== "" && description !== "") {
+        if (country_id && name !== "" && description !== "") {
             setFormInputClass(CLASSES.FORM_INPUT_SUCCESS);
-        } else if (countryId || name !== "" || description !== "") {
+        } else if (country_id || name !== "" || description !== "") {
             setFormInputClass(CLASSES.FORM_INPUT_DEFAULT)
         } else {
             setFormInputClass(CLASSES.FORM_INPUT_ERROR);
         }
-    }, [name, description, countryId, setFormInputClass])
+    }, [name, description, country_id, setFormInputClass])
 
     return (
         <main className={"container-fluid main-container"}>
             <div className={"row row-padding--main"}>
                 <div className={"col-12"}>
-                    <AdminH1 text={LABELS_AND_HEADINGS.ADD_PUBLISHER}/>
+                    <HeadingWithBreadCrumbs text={LABELS_AND_HEADINGS.ADD_PUBLISHER}/>
                 </div>
             </div>
             <div className={"row row-padding--secondary"}>
@@ -71,7 +70,7 @@ export const AdminPublisherAdd = () => {
                             <select
                                 id="country"
                                 className={formInputClass}
-                                onChange={(e) => setCountryId(e.target.value)}>
+                                onChange={(e) => setCountry_id(e.target.value)}>
                                 <option value={""}>{LABELS_AND_HEADINGS.CHOOSE}</option>
                                 {printOptions(countryData)}
                             </select>
@@ -80,9 +79,9 @@ export const AdminPublisherAdd = () => {
                                 onClick={() => addPublisherData({
                                     name: name,
                                     description: description,
-                                    countryId: countryId
-                                }, setFormMessage).then(() => resetAddPublisherForm())}
-                                disabled={!countryId || name === "" || description === ""}>
+                                    country_id: country_id
+                                }, setInformationMessage).then(() => resetAddPublisherForm())}
+                                disabled={!country_id || name === "" || description === ""}>
                             {LABELS_AND_HEADINGS.ADD}
                         </button>
                         <button className={"btn btn-outline-secondary"}
@@ -90,12 +89,6 @@ export const AdminPublisherAdd = () => {
                             {LABELS_AND_HEADINGS.RESET_FORM}
                         </button>
                         <ArrowLeftButton onClick={() => handleBacking(navigate)} label={LABELS_AND_HEADINGS.BACK}/>
-                        {
-                            formMessage.show &&
-                            <p className={formMessage.error ? "alert alert-danger mt-3" : "alert alert-success mt-3"}>
-                                {formMessage.message}
-                            </p>
-                        }
                     </div>
                 </div>
             </div>
